requirements.txt
typer>=0.12.0
amazon-photos>=0.0.97
rich>=13.0.0
pyarrow>=14.0.0
msal>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0
piexif>=1.1.3
Pillow>=10.0.0
mutagen>=1.47.0
# HEIC/HEIF: pip install pillow-heif (recommended for iPhone photos)
